Rename src subdirectory to pkg
This makes directory naming consistent with Fleet.
